--- a/src/Slugify.Core/SlugHelper.cs
+++ b/src/Slugify.Core/SlugHelper.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
+using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Xml;
@@ -36,7 +37,13 @@ namespace Slugify
             inputString = CleanWhiteSpace(inputString, _config.CollapseWhiteSpace);
             inputString = ApplyReplacements(inputString, _config.StringReplacements);
             inputString = RemoveDiacritics(inputString);
-            inputString = DeleteCharacters(inputString, _config.DeniedCharactersRegex);
+
+            string regex = _config.DeniedCharactersRegex;
+            if (regex == null)
+            {
+                regex = "[^" + Regex.Escape(string.Join("", _config.AllowedChars)).Replace("-", "\\-") + "]";
+            }
+            inputString = DeleteCharacters(inputString, regex);
 
             if (_config.CollapseDashes)
             {
@@ -91,14 +98,33 @@ namespace Slugify
         /// </summary>
         public class Config
         {
+            // TODO: Implement a source generator so this can be done at compile time :)
+            private static readonly char[] s_allowedChars =
+                ("abcdefghijklmnopqrstuvwxyz" +
+                "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
+                "0123456789" +
+                "-._").ToCharArray();
+
+            private readonly HashSet<char> _allowedChars = new HashSet<char>(s_allowedChars);
+
             public Dictionary<string, string> StringReplacements { get; set; } = new Dictionary<string, string>
                 {
                     { " ", "-" }
                 };
-            
+
             public bool ForceLowerCase { get; set; } = true;
             public bool CollapseWhiteSpace { get; set; } = true;
-            public string DeniedCharactersRegex { get; set; } = @"[^a-zA-Z0-9\-\._]";
+            /// <summary>
+            /// Note: Setting this property will stop the AllowedChars feature from being used
+            /// </summary>
+            public string DeniedCharactersRegex { get; set; }
+            public HashSet<char> AllowedChars
+            {
+                get
+                {
+                    return DeniedCharactersRegex == null ? _allowedChars : throw new InvalidOperationException("After setting DeniedCharactersRegex the AllowedChars feature cannot be used.");
+                }
+            }
             public bool CollapseDashes { get; set; } = true;
             public bool TrimWhitespace { get; set; } = true;
         }