I'm done. Removed all of the regexes, except a backwards compatible mode for people using DeniedCharacters. The rest is (mostly) allocation free. There is more that could be done but I think this is a nice balance where its mostly readable, but also pretty good.

Fully backwards compatible (I think!) and a reasonable improvement IMO.

On my laptop, on battery:

|   Method |      Mean |    Error |   StdDev | Ratio | RatioSD |      Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|---------:|---------:|------:|--------:|-----------:|------:|------:|----------:|
| Baseline | 116.33 ms | 2.318 ms | 3.095 ms |  1.00 |    0.00 | 22200.0000 |     - |     - |  44.66 MB |
| Improved |  53.87 ms | 1.072 ms | 1.669 ms |  0.46 |    0.02 |  2666.6667 |     - |     - |   5.34 MB |