@ctolkien 2020-05-04
Resolves #5